magnet/requirements.txt
torch
sentence-transformers
pandas
tqdm
spacy
transformers
faiss-cpu
boto3
nats-py